+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.22.0] - 2025-05-19
+### Added
+- View wallet transactions in a bottom sheet when selecting an account.
+- Edit or delete an account from the bottom sheet header.
+
 ## [0.21.0] - 2025-05-19
 ### Added
 - Progress bar for category budgets.